added dev container and created checker structs
