+++ .devcontainer/Dockerfile
@@ -1,0 +1,9 @@
+FROM ubuntu:18.04
+
+WORKDIR /home/
+
+COPY . .
+
+RUN bash ./setup.sh
+
+ENV PATH="/root/.cargo/bin:$PATH"
