+++ .devcontainer/devcontainer.json
@@ -1,0 +1,23 @@
+{
+	"name": "Codespaces Rust Starter",
+	"customizations": {
+        "vscode": {
+            "extensions": [
+                "cschleiden.vscode-github-actions",
+                "ms-vsliveshare.vsliveshare",
+                "matklad.rust-analyzer",
+                "serayuzgur.crates",
+                "vadimcn.vscode-lldb"
+            ],
+            "settings": {
+                "editor.formatOnSave": true,
+                "terminal.integrated.profiles.linux": "/usr/bin/zsh",
+                "files.exclude": {
+                    "**/CODE_OF_CONDUCT.md": true,
+                    "**/LICENSE": true
+                }
+            }
+        }
+    },
+	"dockerFile": "Dockerfile"
+}
